requirements.txt
fastapi
uvicorn
pdfminer.six
ifcopenshell
